fix: comprehensive security and quality improvements
- Remove public API routes without authentication
- Remove debug routes from production
- Add missing Dockerfile and Apache configuration
- Fix database configuration (MySQL as default)
- Add PHPStan configuration for static analysis
- Pin wildcard dependencies to specific versions
- Fix placeholder performance tests with real assertions
- Remove legacy CI files
- Fix code style issues with Laravel Pint
- Add PWA icon generation script

All critical security issues resolved

+++ app/Services/AI/ContinuousQualityMonitor.php
@@ -262,7 +262,7 @@ class ContinuousQualityMonitor
             $memoryUsage = (int) $matches[1];
             $maxMemory = 512;
 
-            return max(0, 100 - (($memoryUsage / $maxMemory) * 100));
+            return (int) max(0, 100 - (($memoryUsage / $maxMemory) * 100));
         }
 
         return 100;
@@ -357,8 +357,8 @@ class ContinuousQualityMonitor
      */
     public function getAlertsSummary(): array
     {
-        $criticalAlerts = array_filter($this->alerts, fn($alert): bool => is_array($alert) && is_string($alert['type'] ?? null) && $alert['type'] === 'critical');
-        $warningAlerts = array_filter($this->alerts, fn($alert): bool => is_array($alert) && is_string($alert['type'] ?? null) && $alert['type'] === 'warning');
+        $criticalAlerts = array_filter($this->alerts, fn ($alert): bool => is_array($alert) && is_string($alert['type'] ?? null) && $alert['type'] === 'critical');
+        $warningAlerts = array_filter($this->alerts, fn ($alert): bool => is_array($alert) && is_string($alert['type'] ?? null) && $alert['type'] === 'warning');
 
         return [
             'total' => count($this->alerts),
